A Sieve of Eratosthenes class
This simple class allows for the creation of a Sieve of Eratosthenes
which constructs a boolean array that stores if a particular value (or
value + an offset if a range of values is given) is composite. The user
can then ask if a particular value within the range is prime or for all
the primes in sieve as well as the number of primes in the sieve.