Indexed Binary Search Tree
It is based on an implementation I of a standard BST I found online at
cprograming.com (as of 1/14/14 it is lesson 18) with a few
modifications based on my personal taste and what the assignment called
for.

Feel free to use it/if you see any issues, let me know so I can fix it
before I turn it in.